moved Checkpoint.java to controller
